Moving createCard to ListOperations
